# docs/swagger.yaml
definitions:
  errorcode.Error:
    type: object
  model.Tag:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      state:
        type: integer
    type: object
info:
  contact: {}
  description: 部落格建立
  termsOfService: https://github.com/go-programing-tour-book
  title: 部落格系統
  version: "1.0"
paths:
  /api/v1/tags:
    get:
      parameters:
      - description: 標籤名稱
        in: query
        name: name
        type: string
      - description: 狀態
        in: query
        name: state
        type: integer
      - description: 頁碼
        in: query
        name: page
        type: integer
      - description: 每頁數量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 請求錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 內部錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 取得多個標籤
    post:
      parameters:
      - description: 標籤名稱
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: 狀態
        in: body
        name: state
        schema:
          type: integer
      - description: 建立者
        in: body
        name: created_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 請求錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 內部錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 新增標籤
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: 標籤ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 請求錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 內部錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 刪除標籤
    put:
      parameters:
      - description: 標籤ID
        in: path
        name: id
        required: true
        type: integer
      - description: 標籤名稱
        in: body
        name: name
        schema:
          type: string
      - description: 狀態
        in: body
        name: state
        schema:
          type: integer
      - description: 修改者
        in: body
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 請求錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 內部錯誤
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 更新標籤
swagger: "2.0"